Added the project for DataLibrary
Added the Library Class project called DataLibrary to the solution to separate the data access (mysql and dapper) from the backend part for better separate of concerns.

Modifies the test controller to test a small query instead of a single string.

Added the admin_user class to the modules.

Changed AuthController class to AccountController for better matching the class diagram.

Was working on the AccountController.