Refactoring some parts in the backend
Modified the entities (user, student, instructor, admin) in the account management class diagram to better match the ERD model.
Modified the models in the backend to better match the class diagram.

Refactoring some parts in the account controller class in the backend, we removed the logUser model and substituted it in the functions with the json element.