Removed the logout from the account management class
We removed the logout function from the class since it will not be implemented in the backend.
The backend does not store the token generated when the user log in, so we have no way to invalidate it, it will just expire with time.
The frontend will just remove the token when the user says logout.
The system assumes that the user has the right to access it from different devices/interfaces at the same time.