Added the department management part.
Modified the SQL files to include the departments and department priority list.
Added the controller for handelling department stuff.
Added the models :Department, StudentDepartmentPrioirty.
Added the main 4 functions of the department controller.
The functions is just prototype but the specification is valid.
Modified the diagrams to match the changes.
No haed of department will be added to the department.
the process of assigning departments to student is manual done by the admin.