Modified the courses in the database
We divided the course table into three tables, course, course instance, course department applicability.
Added the course prerequisite table to represent the courses required for a specific course.
We modified the construction and drop scripts.
We hashed the passwords stored into the database.
We added Hashing Tester console project to test and generate the hash corresponding to a specific password.
Small modification to the priority list document.