Add one more functions and some changes
We added the add course instance function to the backend.
We extracted the term from the course and made it a stand-alone class in its own file, now it is called term.
We added Time Utilities class which have some function related to time like get current year, get current term, etc...