Massive Revision to the backend functions.
Added more checks to the functions.
Modified and simplified the flow of some functions.
Added message to the return message whether it is a bad request or ok.
Remove the URL parameters from the patch functions in the profile controller.
Modified the names of some parameters.
Added student course registration to the SQL script (still in work).
The backend API is bound to change if we managed to increase the benefit of the token like removing the explicit id and user type from get profile and save profile changes functions.
Massive formatting and refactoring.